[int] Use 32-bit flags in ‘surface-color-key!’.
This is a bug but not marked "fix bug" because pre-release,
from original commit, caught luckily by inspection -- at this
time there is no straightforward way to determine a surface's
color key if any.

* src/sdlsurface.c (surface_color_key_x): Use ‘Uint32’ for ‘cflags’.

+++ src/sdlsurface.c
@@ -311,7 +311,7 @@ Return @code{#t} if successful.  */)
 {
 #define FUNC_NAME s_surface_color_key_x
   Uint32 cpixel;
-  Uint8 cflags;
+  Uint32 cflags;
 
   ASSERT_SURFACE (surface, 1);
   if (EXACTLY_FALSEP (pixel))
